flatten {mzID}	R Documentation

Flatten an mzID related class into a table

Description

This function flattens the content of the object into a table by merging the content intelligently (it knows the links between the different objects).

Usage

flatten(object, safeNames = TRUE)

## S4 method for signature 'mzIDpsm'
flatten(object, safeNames = TRUE)

## S4 method for signature 'mzIDpeptides'
flatten(object, safeNames = TRUE)

## S4 method for signature 'mzID'
flatten(object, safeNames = TRUE)

## S4 method for signature 'mzIDCollection'
flatten(object, safeNames = TRUE)

Arguments

object
The object to be flattened

safeNames
Logical. Should column names be lowered to ensure compitability between different versions of the mzIdentML schema. Defaults to TRUE

Value

A data.frame with the flattened result or a list of data.frames

Methods (by class)

mzIDpsm: Merge id and scans according to the mapping

mzIDpeptides: Merge peptides with their modifications

mzID: Flatten an mzID object with respect to psm'

mzIDCollection: Flatten all mzID object in the collection into a list of data frames.

See Also

mzID-class mzIDCollection-class mzIDpsm mzIDpeptides

Examples

exampleFile <- system.file('extdata', '55merge_tandem.mzid', package = 'mzID')
mzResults <- mzID(exampleFile)
head(flatten(mzResults))

[Package mzID version 1.22.0 Index]