increment {mzID}	R Documentation

Tools to handle generation of lookup names for the dictionary

Description

increment looks for a counter in the data environment of mzIDCollection objects and increments it by one if it exists. Otherwise it initialises the counter to 1. It returns the value of the counter.

Usage

increment(object)

## S4 method for signature 'mzIDCollection'
increment(object)

Arguments

object
An mzIDCollection object

Value

An integer given the current count

Methods (by class)

mzIDCollection: Increment the counter in mzIDcollection objects

[Package mzID version 1.22.0 Index]