keyFor {mzID}	R Documentation

Tools to handle generation of lookup names for the dictionary

Description

keyFor returns the internal hash for a given name.

Usage

keyFor(object, name)

## S4 method for signature 'mzIDCollection,character'
keyFor(object, name)

Arguments

object
An mzIDCollection object

Value

The hash under wich the object with the given name is stored

Methods (by class)

object = mzIDCollection,name = character: Get the key for sample in an mzIDCollection object

[Package mzID version 1.24.0 Index]