mzID-class {mzID}	R Documentation

A class to contain data from mzIdentML-files

Description

This class stores all parsed information from mzIdentML files

Usage

## S4 method for signature 'mzID'
show(object)

## S4 method for signature 'mzID'
length(x)

## S4 method for signature 'mzID'
removeDecoy(object)

## S4 method for signature 'mzID'
database(object, safeNames = TRUE)

## S4 method for signature 'mzID'
evidence(object, safeNames = TRUE)

## S4 method for signature 'mzID'
parameters(object)

## S4 method for signature 'mzID'
software(object)

## S4 method for signature 'mzID'
files(object)

## S4 method for signature 'mzID'
peptides(object, safeNames = TRUE)

## S4 method for signature 'mzID'
modifications(object)

## S4 method for signature 'mzID'
id(object, safeNames = TRUE)

## S4 method for signature 'mzID'
scans(object, safeNames = TRUE)

## S4 method for signature 'mzID'
idScanMap(object)

## S4 method for signature 'mzID'
c(x, y, ..., recursive = FALSE)

Arguments

object
An mzID object

x
An mzID object

safeNames
Should column names be lowercased to ensure compatibility between v1.0 and v1.1 files?

y
An mzID or mzIDCollection object

...
ignored

recursive
ignored

Details

The mzID class stores information in a subset of classes, each class having its own slot. While these classes should not need to be accessed directly, descriptions of their content is delegated to each respective class.

Methods (by generic)

show: Short summary of object content

length: Get number of psm' in object

removeDecoy: Remove decoys from mzID object

database: Get the database used for searching

evidence: Get the evidence from the peptide search

parameters: Get the parameters used for the search

software: Get the software used to arrive at the results

files: Get the data files used for the analysis

peptides: Get the peptides identified.

modifications: Get the modification on the identified peptides

id: Get the identification results

scans: Get the scans matched to peptides

idScanMap: Get the link between scans and identifications

c: Combine mzID and mzIDCollection objects

Slots

parameters

An instance of mzIDparameters-class. This object contains all information related to how the analysis was carried out.

psm

An instance of mzIDpsm-class. This object contains the meat of the analysis with all scans and their related PSMs recorded.

peptides

An instance of mzIDpeptides-class. This object contains a library of all peptides generated from the database along with possible modifications.

evidence

An instance of mzIDevidence-class. This object lists all peptides detected in the analysis with reference to the mzIDpeptides instance.

database

An instance of mzIDdatabase-class. This object contains information on the proteins in the database. As the full database is not recorded in mzIdentML files the actual protein sequence is not recorded but there is sufficient information to retrieve it from the database file.

Objects from the class

Objects can be created using the mzID constructor, which handles parsing of mzIdentML files

References

http://www.psidev.info/mzidentml

See Also

mzID

Other mzID.classes: mzIDCollection-class, mzIDdatabase-class, mzIDevidence-class, mzIDparameters-class, mzIDpeptides-class, mzIDpsm-class

[Package mzID version 1.24.0 Index]