mzID-getters {mzID}	R Documentation

Getter functions for identification data

Description

This set of functions are used to extract data from mzID and mzIDCollection objects.

Usage

evidence(object, safeNames = TRUE)

id(object, safeNames = TRUE)

idScanMap(object)

parameters(object)

software(object)

files(object)

Arguments

object
An mzID or mzIDCollection object

safeNames
Logical. Should column names be lowered to ensure compitability between different versions of the mzIdentML schema. Defaults to TRUE

Value

A data frame or a list of data frames in the case of mzIDCollections

See Also

mzID-class mzIDCollection-class

[Package mzID version 1.22.0 Index]