mzID-package {mzID}	R Documentation

A parser for the mzIdentML file format

Description

Using the mzID function this package is able to parse mzIdentML files into an mzID class instance. Multiple files can be parsed in parallel into an mzIDCollection which is a list-like class that handles multiple mzID objects.

Details

The key functionalities are described in the following pages:

mzID:Parse mzIdentML files

mzID-class:Class to store the data from an mzIdentML file

mzIDCollection-class:Class to store multiple mzID objects

mzID-getters:Access the data stored in mzID and mzIDCollection objects

flatten:Converts the content of an mzID object to a table with respect to the peptide-spectrum matches

The package is maintained at its GitHub repository where feature requests and bugs can be directed. Questions and bugs can furthermore be posted at the Bioconductor support site

Author(s)

Thomas Lin Pedersen with contributions from Laurent Gatto

[Package mzID version 1.22.0 Index]