mzIDCollection-class {mzID}	R Documentation

A class to handle a set of mzID objects

Description

This class is a container for multiple mzID objects. It is constructed such that the bulk data are not copied when passed around. It is the aim that this class have parity with the mzID class in the methods it exposes to the user, such that mzIDCollections can be thought of as vectors in the traditional R sense. Furthermore it accepts standard indexing and concatenation.

Usage

## S4 method for signature 'mzIDCollection'
show(object)

## S4 method for signature 'mzIDCollection'
length(x)

as.list.mzIDCollection(object)

## S4 method for signature 'mzIDCollection'
removeDecoy(object)

## S4 method for signature 'mzIDCollection'
names(x)

## S4 replacement method for signature 'mzIDCollection,character'
names(x) <- value

## S4 method for signature 'mzIDCollection,numeric,missing'
x[[i, j, ...]]

## S4 method for signature 'mzIDCollection,character,missing'
x[[i, j, ...]]

## S4 method for signature 'mzIDCollection,numeric,missing,missing'
x[i, j, drop]

## S4 method for signature 'mzIDCollection,character,missing,missing'
x[i, j, drop]

## S4 method for signature 'mzIDCollection,logical,missing,missing'
x[i, j, drop]

## S4 method for signature 'mzIDCollection'
c(x, y, ..., recursive = FALSE)

## S4 method for signature 'mzIDCollection'
database(object, safeNames = TRUE)

## S4 method for signature 'mzIDCollection'
evidence(object, safeNames = TRUE)

## S4 method for signature 'mzIDCollection'
parameters(object)

## S4 method for signature 'mzIDCollection'
software(object)

## S4 method for signature 'mzIDCollection'
files(object)

## S4 method for signature 'mzIDCollection'
peptides(object, safeNames = TRUE)

## S4 method for signature 'mzIDCollection'
modifications(object)

## S4 method for signature 'mzIDCollection'
id(object, safeNames = TRUE)

## S4 method for signature 'mzIDCollection'
scans(object, safeNames = TRUE)

## S4 method for signature 'mzIDCollection'
idScanMap(object)

Arguments

object
An mzIDCollection object

x
An mzIDCollection object

value
A character vector of desired names

i
An integer or a string giving the index or the name respectively

j
ignored

...
ignored

drop
ignored

y
An mzID or mzIDCollection object

recursive
ignored

safeNames
Should column names be lowercased to ensure compatibility between v1.0 and v1.1 files?

Details

Objects of this class is usually constructed be passing mulitple files to the mzID constructor, or by combining mulitple mzID objects.

Methods (by generic)

show: A short summary of the content of the object

length: Return the number of mzID object in the collection

removeDecoy: Removes decoys in all mzID object in collection

names: Get the names of the mzID object stored in the collection

names<-: Set the names of the mzID object stored in the collection

[[: Extract an mzID object by index

[[: Extract an mzID object by name

[: Subset collection by index

[: Subset collection by name

[: Subset collection by logical value

c: Combine mzIDCollction and mzID objects

database: Get the database used for searching

evidence: Get the evidence from the peptide search

parameters: Get the parameters used for the search

software: Get the software used to arrive at the results

files: Get the data files used for the analysis

peptides: Get the peptides identified.

modifications: Get the modification on the identified peptides

id: Get the identification results

scans: Get the scans matched to peptides

idScanMap: Get the link between scans and identifications

Slots

data

An environment that holds the individual mzID objects

.lookup

A matrix with indexing information for retriving the mzID objects in the @data slot.

See Also

mzID mzIDCollection

Other mzID.classes: mzID-class, mzIDdatabase-class, mzIDevidence-class, mzIDparameters-class, mzIDpeptides-class, mzIDpsm-class

[Package mzID version 1.22.0 Index]