mzIDdatabase-class {mzID}	R Documentation

A class to store database information from an mzIdentML file

Description

This class handles parsing and storage of database information from mzIDentML files, residing at the /MzIdentML/SequenceCollection/DBSequence node.

Usage

## S4 method for signature 'mzIDdatabase'
show(object)

## S4 method for signature 'mzIDdatabase'
length(x)

## S4 method for signature 'mzIDdatabase'
database(object, safeNames = TRUE)

Arguments

object
An mzIDevidence object

x
An mzIDdatabase object

safeNames
Should column names be lowercased to ensure compatibility between v1.0 and v1.1 files?

Details

The content of the class is stored in a data.frame with columns depending on the content of the mzIdentML file. Required information for files conforming to the mzIdentML standard are: 'accession', 'searchDatabase_ref' and 'id', while additional information can fx be 'length' (number of residues), 'description' (from the fasta file) and 'sequence' (the actual sequence).

Methods (by generic)

show: Short summary of the content of the object

length: Report the number of proteins in the database

database: Get the database used for searching

Slots

database

A data.frame containing references to all the database sequences from the mzIdentML file

Objects from the class

Objects of mzIDdatabase are not meant to be created explicitly but as part of the mzID-class. Still object can be created with the constructor mzIDdatabase.

See Also

mzIDdatabase

Other mzID.classes: mzID-class, mzIDCollection-class, mzIDevidence-class, mzIDparameters-class, mzIDpeptides-class, mzIDpsm-class

[Package mzID version 1.22.0 Index]