mzIDevidence {mzID}	R Documentation

A constructor for the mzIDevidence class

Description

This function handles parsing of data and construction of an mzIDevidence object. This function is not intended to be called explicitly but as part of an mzID construction. Thus, the function is not exported.

Usage

mzIDevidence(doc, ns, addFinalizer = FALSE, path)

Arguments

doc
an XMLInternalDocument created using xmlInternalTreeParse

ns
The appropriate namespace for the doc, as a named character vector with the namespace named x

addFinalizer
Logical Sets whether reference counting should be turned on

path
If doc is missing the file specified here will be parsed

Value

An mzIDevidence object

See Also

mzIDevidence-class

[Package mzID version 1.22.0 Index]