mzIDevidence-class {mzID}	R Documentation

A class to store peptide evidence information from an mzIdentML file

Description

This class handles parsing and storage of peptide evidence information from mzIDentML files, residing at the /*/x:SequenceCollection/x:PeptideEvidence node.

Usage

## S4 method for signature 'mzIDevidence'
show(object)

## S4 method for signature 'mzIDevidence'
length(x)

## S4 method for signature 'mzIDevidence'
evidence(object, safeNames = TRUE)

Arguments

object
An mzIDevidence object

x
An mzIDevidence object

safeNames
Should column names be lowercased to ensure compatibility between v1.0 and v1.1 files?

Details

The content of the class is stored in a data.frame with columns depending on the content of the mzIdentML file. Columns represent the attribute values of for each PeptideEvidence node. For files conforming to the HUPO standard, dbSequence_ref, id and peptide_ref is required while start, end, pre, post, name, isDecoy, frame and translationTable_ref are optional. Information residing in cvParam and userParam children are not parsed.

Methods (by generic)

show: Short summary of the content of the object

length: Report number of evidence

evidence: Get the evidence from the peptide search

Slots

evidence

A data.frame containing all peptide evidence from the mzIdentML file

Objects from the class

Objects of mzIDevidence are not meant to be created explicitly but as part of the mzID-class. Still object can be created with the constructor mzIDevidence.

See Also

mzIDevidence

Other mzID.classes: mzID-class, mzIDCollection-class, mzIDdatabase-class, mzIDparameters-class, mzIDpeptides-class, mzIDpsm-class

[Package mzID version 1.24.0 Index]