mzIDparameters-class {mzID}	R Documentation

A Class to store analysis information from the mzIdentML file

Description

This class tries to collect the multitude of different analysis information required to rerun the analysis. The intended data to be stored are: The software used in the analysis of the data, the location and nature of the rawfile(s), the location and nature of the database file(s), the location of the mzIDentML file itself as well as all the parameters used during the analysis leading to the mzIdentML file. Information regarding how the LC-MS experiment was performed should be collected from the raw data file. As the parameters used in different software solutions can vary greatly, all these parameters are stored in a named list, which can thus be very different from pipeline to pipeline. It is the users responsibility to check conformity between samples.

Usage

## S4 method for signature 'mzIDparameters'
show(object)

## S4 method for signature 'mzIDparameters'
length(x)

## S4 method for signature 'mzIDparameters'
parameters(object)

## S4 method for signature 'mzIDparameters'
software(object)

## S4 method for signature 'mzIDparameters'
files(object)

Arguments

object
An mzIDparameters object

x
An mzIDparameters object

Methods (by generic)

show: Short summary of the content

length: Get the length of the object

parameters: Get the parameters used for the search

software: Get the software used to arrive at the results

files: Get the data files used for the analysis

Slots

software

A data frame with information retaining to the software used for the analysis. At least the name and an id is given, but optionally also version number and URI.

rawFile

A data frame with information about the raw data file(s) used for the analysis. The data frame will contain at least the location and spectrum ID format.

databaseFile

A data frame containing at least the location and file format of the database file used in the search.

idFile

A character string containing the location of the mzIdentML file at the time of parsing.

parameters

A list containing containing the information stored in the MzIdentML/AnalysisProtocolCollection/SpectrumIdentificationProtocol node. SearchType and Threshold are the only required parameters given by the mzIdentML standard.

Objects from the class

Objects of mzIDparameters are not meant to be created explicitly but as part of the mzID-class. Still object can be created with the constructor mzIDparameters (not exported).

See Also

mzIDparameters

Other mzID.classes: mzID-class, mzIDCollection-class, mzIDdatabase-class, mzIDevidence-class, mzIDpeptides-class, mzIDpsm-class

[Package mzID version 1.22.0 Index]