mzIDpeptides-class {mzID}	R Documentation

A class to store peptide information from an mzIdentML file

Description

This class handles parsing and storage of peptide information from mzIDentML files, residing at the /x:MzIdentML/x:SequenceCollection/x:Peptide node.

Usage

## S4 method for signature 'mzIDpeptides'
show(object)

## S4 method for signature 'mzIDpeptides'
length(x)

## S4 method for signature 'mzIDpeptides'
peptides(object, safeNames = TRUE)

## S4 method for signature 'mzIDpeptides'
modifications(object)

Arguments

object
An mzIDpeptides object

x
An mzIDpeptides object

safeNames
Should column names be lowercased to ensure compatibility between v1.0 and v1.1 files?

Details

The information is stored in a dataframe with an id, an optinal name and the amino acid sequence of the peptide. Alongside a list is stored with modification information of each peptide. Each row in the dataframe has a corresponding entry en the list. If no modification of the peptide is present the entry is NULL, if a modification is present the entry is a dataframe, listing the different modifications of the peptide.

Methods (by generic)

show: Short summary of the content of the object

length: Report the number of peptides

peptides: Get the peptides identified.

modifications: Get the modification on the identified peptides

Slots

peptides

A data.frame containing all peptides used in the search

modifications

A list containing possible modifications of the peptides listed in @peptides

Objects from the class

Objects of mzIDpeptides are not meant to be created explicitly but as part of the mzID-class. Still object can be created with the constructor mzIDpeptides.

See Also

mzIDpeptides

Other mzID.classes: mzID-class, mzIDCollection-class, mzIDdatabase-class, mzIDevidence-class, mzIDparameters-class, mzIDpsm-class

[Package mzID version 1.22.0 Index]