mzIDpsm-class {mzID}	R Documentation

A class to store psm information from an mzIdentML file

Description

This class handles parsing and storage of scan info and the related psm's. This information resides in the /*/x:DataCollection/x:AnalysisData/x:SpectrumIdentificationList/x:SpectrumIdentificationResult node.

Usage

## S4 method for signature 'mzIDpsm'
show(object)

## S4 method for signature 'mzIDpsm'
length(x)

## S4 method for signature 'mzIDpsm'
id(object, safeNames = TRUE)

## S4 method for signature 'mzIDpsm'
scans(object, safeNames = TRUE)

## S4 method for signature 'mzIDpsm'
idScanMap(object)

Arguments

object
An mzIDpsm object

x
An mzIDpsm object

safeNames
Should column names be lowercased to ensure compatibility between v1.0 and v1.1 files?

Details

The content of the class is stored as two data frames: One containing a row for each scan in the results, and one containing all psm's in the results. Additionally a list containing indexing from scan to psm is stored.

Methods (by generic)

show: A short summary of the content

length: Get the number of psm'

id: Get the identification results

scans: Get the scans matched to peptides

idScanMap: Get the link between scans and identifications

Slots

scans

A data.frame containing all reference to all scans with at least one psm. The columns gives at least an ID, a spectrumID and a reference to the file used.

id

A data.frame containing all psm's from the analysis. The columns depend on the file but at least id, chargeState, experimentalMassToCharge, passThreshold and rank must exist according to the mzIdentML specifications.

mapping

A list with an entry for each row in @scans. Each entry contains an integer vector pointing to the related rows in @id.

Objects from the class

Objects of mzIDpsm are not meant to be created explicitly but as part of the mzID-class. Still object can be created with the constructor mzIDpsm.

See Also

mzIDpsm

Other mzID.classes: mzID-class, mzIDCollection-class, mzIDdatabase-class, mzIDevidence-class, mzIDparameters-class, mzIDpeptides-class

[Package mzID version 1.22.0 Index]