removeDecoy {mzID}	R Documentation

Remove decoy identification

Description

This function trims down an mzID or mzIDCollection object by removing all information that is only related to the decoy database search. If some information relates to both the regular and decoy database (e.g. a peptide sequence that can be found in both databases) it is kept.

Usage

removeDecoy(object, ...)

Arguments

object
An mzID or mzIDCollection to remove decoy information from

...
Currently ignored

Value

An mzID or mzIDCollection object depending on the input

See Also

mzID-class mzIDCollection-class

[Package mzID version 1.24.0 Index]